requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
lxml==4.9.3
fake-useragent==1.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.3